// src/models/resume.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export type Resume = {
  $schema: string
  basics: Basics
  work: Work[]
  volunteer: Volunteer[]
  education: Education[]
  awards: Award[]
  //   certificates: any[]
  //   publications: any[]
  skills: Skill[]
  languages: Language[]
  //   interests: any[]
  //   references: any[]
  projects: Project[]
  meta: Meta
}

export interface Award {
  title: string
  date: string
  awarder: string
  summary: string
}

export interface Basics {
  name: string
  label: string
  image: string
  email: string
  phone: string
  url: string
  summary: string
  location: LocationClass
  profiles: Profile[]
}

export interface LocationClass {
  countryCode: string
  address: string
}

export interface Profile {
  network: string
  username: string
  url: string
}

export interface Education {
  institution: string
  area: string
  studyType: string
  startDate: Date
  endDate: Date
  score: string
  courses: string[]
}

export interface Language {
  language: string
  fluency: string
}

export interface Meta {
  version: string
  canonical: string
}

export interface Project {
  name: string
  startDate: Date
  summary: string
  url: string
  endDate: Date
}

export interface Skill {
  name: string
  level: string
  //   keywords: any[]
}

export interface Volunteer {
  organization: string
  position: string
  startDate: Date
  endDate: Date
  summary?: null | string
  //   highlights: any[]
  url: string
}

export interface Work {
  position: string
  startDate: Date
  endDate: Date
  summary?: null | string
  //   highlights: any[]
  url: string
  name: string
  location?: string
}

export enum LocationEnum {
  SãoPauloAreaBrazil = 'São Paulo Area, Brazil',
  SãoPauloBrazil = 'São Paulo, Brazil',
}